Moved API keys to .env file to be able to add moduleConfig.json to Git
Moved module config from config.json to own moduleConfig.json file
Renamed server_config.json to serverConfig.json
Adjusted copypasta strings
Added more error messages to imagetext module
Updated dependencies
