--- config_default.json
@@ -1,3 +1,0 @@
-{
-	"apikey": "put_your_api_key_here"
-}
