+++ config_example.json
@@ -1,0 +1,3 @@
+{
+	"superUsers": []
+}
