--- embed_example.json
@@ -1,29 +1,0 @@
-{
-	"embed": {
-		"author": {
-			"name": "client.user.username",
-			"icon_url": "client.user.avatarURL"
-		},
-		"title": "**guild.name**",
-		"thumbnail": "guild.iconURL",
-		"fields": [
-			{
-				"name": "ID",
-				"value": "guild.id"
-			},
-			{
-				"name": "Members",
-				"value": "guild.memberCount"
-			},
-			{
-				"name": "Joined when",
-				"value": "guild.joinedAt"
-			}
-		],
-		"timestamp": "new Date()",
-		"footer": {
-			"icon_url": "client.user.avatarURL",
-			"text": "This has been an uwubot service"
-		}
-	}
-}
