+++ example.env
@@ -1,0 +1,5 @@
+API_KEY="Discord API key"
+
+# The Flickr API keys are necessary to use the "animals" module
+FLICKR_API="Flickr API key"
+FLICKR_SECRET="Flickr API secret"
