+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "uwubot",
-  "version": "1.2.3",
+  "version": "1.3.0",
   "description": "A meme bot for spenglers",
   "main": "bot.js",
   "dependencies": {
     "bufferutil": "latest",
     "console-stamp": "^0.2.9",
-    "cross-spawn": "^7.0.1",
+    "cross-spawn": "^7.0.2",
     "deepmerge": "^4.2.2",
-    "discord.js": "^11.5.1",
+    "discord.js": "^11.6.4",
+    "dotenv": "^8.2.0",
     "emoji-aware": "^3.0.5",
     "flickrapi": "^0.6.2",
     "tmp": "^0.1.0"
